Clarify where to create the new git repo.
Be more explicit about where the new repo should be located. It's much
easier to set up a repo on the internet. This was implied before and now
it's explicit.

The first time I helped a colleague set this up, we used a local repo.
Since we didn't know what we were doing, we screwed it up pretty badly.
This bit of text might allow others not to make the same mistake we did.